Expose feedback prompt globally and refresh QR balance
