+++ plan.md
@@ -84,6 +84,8 @@
 - [x] Главный экран: карточка баланса/уровня/кэшбэка обновлена (белые подписи, кнопка «Нажмите» без подчёркивания, стрелка «Назад» убрана).
 - [x] Страница QR: удалена ссылка «Вернуться в профиль», уменьшена подложка вокруг кода, таймер остался рядом с кнопкой.
 - [x] Страница QR: добавлен адаптивный прогресс-бар «Сумма покупок до следующего уровня» с отображением текущей суммы и порога (заглушка).
+- [x] Миниаппа: логика окна отзыва вынесена в общий контроллер, окно появляется на любой странице после покупки.
+- [x] Страница QR автоматически обновляет баланс и уровень по событиям и при возврате в приложение.
 
 - [x] Реферальная программа: поправил типизацию под Prisma v6 (используем `ReferralProgram` из клиента и корректно формируем payload create/update), чтобы `tsc` не падал на сервисе.
 - [x] PortalModule импортирует ReferralModule, чтобы `ReferralService` в портале резолвился без ошибок DI.
